Remove internal libs from tests
